Rename class AdaptiveMedia*ConfigurationTest to AM*ConfigurationTest
