+++ adaptive-media-image-impl-test/src/testIntegration/java/com/liferay/adaptive/media/image/internal/test/BaseAMImageConfigurationTestCase.java
@@ -43,7 +43,7 @@ import org.junit.BeforeClass;
 /**
  * @author Alejandro Hernández
  */
-public abstract class BaseAdaptiveMediaImageConfigurationTestCase {
+public abstract class BaseAMImageConfigurationTestCase {
 
 	@BeforeClass
 	public static void setUpClass() {
